refactor: expiration을 int -> long으로 변경

+++ backend/src/main/java/middle_point_search/backend/common/security/jwt/constants/JwtProperties.java
@@ -27,14 +27,14 @@ public class JwtProperties {
 	@Getter
 	@RequiredArgsConstructor
 	public static class Access {
-		private int expiration;
+		private long expiration;
 		private String header;
 	}
 
 	@Getter
 	@RequiredArgsConstructor
 	public static class Refresh {
-		private int expiration;
+		private long expiration;
 		private String header;
 	}
 }
